Started adding the help instructions to the faq
Gotta go, so I'm making an emergency save-progress commit.
I am also thinking it would be worthwhile to open an issue first
before filing a pull request. It could be the case that these
instructions should be located somewhere else, or that it's better to
take a different approach and avoid the problem entirely from the beginning.

Signed-off-by: WizardPlatypus <f.volik.m@gmail.com>